Advert model and repository, JDK version update, user DTO update

+++ repository/src/main/java/com/project/nairon/repository/naironuser/NaironUserRepository.java
@@ -2,7 +2,6 @@ package com.project.nairon.repository.naironuser;
 
 
 import com.project.nairon.models.naironuser.NaironUser;
-import com.project.nairon.models.naironuser.NaironUserId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,5 +10,5 @@ import org.springframework.stereotype.Repository;
  */
 
 @Repository
-public interface NaironUserRepository extends JpaRepository<NaironUser, Long> {
+public interface NaironUserRepository extends JpaRepository<NaironUser, Integer> {
 }
